// src/app/features/animal-health/vaccination/models/vacc-details.model.ts
export interface VaccinationDetails {
    animalId:            number;
    taggingDate:         string;
    tagId:               number;
    vaccinationType:     string;
    lastDewormingDate:   null;
    lastVaccinationDate: string;
    sex:                 string;
    animalAge:           string;
    vaccinationTypeDesc: string;
    speciesCd:           number;
    speciesName:         string;
    routeCd:             number;
    routeName:           string;
    formCd:              number;
    formName:            string;
    dosage:              number;
    unitCd:              number;
    unitDesc:            string;
}